# contracts/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: Fintcart API Gateway (borde REST)
  version: 1.0.0
  description: >
    Único punto de entrada REST/HTTP del sistema (Constitución Principio II).
    Traduce HTTP/REST ↔ gRPC interno. Aplica CORS, rate limiting (Redis) y
    validación de tokens JWT (blacklist Redis) en cada solicitud.
    PRECISIÓN: todo monto/tasa se serializa como `string` decimal canónica
    (Principio VIII). PROHIBIDO número JSON para dinero.
servers:
  - url: https://api.fintcart.co/v1
components:
  securitySchemes:
    oauth2PKCE:
      type: oauth2
      description: >
        OAuth2 Authorization Code + PKCE (cliente público SPA).
        Los dos endpoints los sirve ESTE Gateway y no un host aparte: el
        Principio II reserva toda la superficie REST al borde y el Servidor de
        Autenticación solo expone gRPC. Ver `/oauth/authorize` y `/oauth/token`
        en `paths:`, y la desviación documentada respecto de RFC 6749 §3.1
        (el endpoint de autorización es JSON, no una redirección de navegador,
        porque no existe ningún componente que pueda servir la página de login).
      flows:
        authorizationCode:
          authorizationUrl: https://api.fintcart.co/v1/oauth/authorize
          tokenUrl: https://api.fintcart.co/v1/oauth/token
          scopes:
            openid: Identidad básica
            learning: Acceso al módulo de aprendizaje
            simulator: Acceso a simuladores
            profile: Gestión de perfil
  schemas:
    DecimalString:
      type: string
      pattern: '^-?\d+(\.\d+)?$'
      example: "1500000.00"
      description: Decimal canónico como string (nunca número JSON).
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
...
    Article:
      type: object
      properties:
        article_id: { type: string, format: uuid }
        title: { type: string }
        category: { type: string }
        body: { type: string }
        current_version_no: { type: integer }
    QuizGradeResult:
      type: object
      properties:
        attempt_id: { type: string, format: uuid }
        attempt_no: { type: integer }
        score: { $ref: '#/components/schemas/DecimalString' }
        passed: { type: boolean }
        points_after: { type: integer }
...
security:
  - oauth2PKCE: []
paths:
...
  /oauth/authorize:
    post:
      summary: Autenticar y emitir un authorization_code ligado al code_challenge (PKCE)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, client_id, redirect_uri, code_challenge, code_challenge_method]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
                client_id: { type: string }
                redirect_uri: { type: string, format: uri }
                code_challenge: { type: string }
                code_challenge_method: { type: string, enum: [S256] }
                scopes: { type: array, items: { type: string } }
      responses:
        '200':
          description: Código de un solo uso
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
                  redirect_uri: { type: string, format: uri }
        '400': { description: PKCE ausente o `code_challenge_method` distinto de S256 }
        '401': { description: Credenciales inválidas (mismo mensaje que cuenta inexistente) }
        '403': { description: Falta verificar el correo (FR-002) }
  /oauth/token:
    post:
      summary: Canjear authorization_code + code_verifier, o rotar un refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type: { type: string, enum: [authorization_code, refresh_token] }
                code: { type: string }
                code_verifier: { type: string }
                client_id: { type: string }
                redirect_uri: { type: string, format: uri }
                refresh_token: { type: string }
      responses:
        '200':
          description: Par de tokens; el refresh SIEMPRE se rota (research D-05)
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  token_type: { type: string, enum: [Bearer] }
                  # Segundos enteros: es el único número JSON del borde, y se
                  # permite porque no es un monto ni una tasa (Principio VIII).
                  expires_in: { type: integer }
        '400': { description: grant_type no soportado o parámetros incompletos }
        '401': { description: Código o refresh token inválido, caducado o reutilizado }

  # --- Identidad (proxy hacia Autenticación / Saga de registro en Orquestador) ---
  /auth/register:
    post:
      summary: Registro de usuario (inicia Saga de registro)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
                display_name: { type: string }
      responses:
        '202': { description: Registro aceptado; correo de verificación enviado }
        '409': { description: Correo ya registrado, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/verify-email:
    post:
      summary: Verificación de correo (inicia Saga de verificación)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, verification_token]
              properties:
                user_id: { type: string, format: uuid }
                verification_token: { type: string }
      responses:
        '200': { description: Correo verificado }
        '400': { description: Token inválido o expirado }
...
  /catalog/articles:
    get:
      summary: Catálogo de artículos publicados (FR-010)
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: page_token, in: query, schema: { type: string } }
      responses:
        '200':
          description: Lista paginada
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Article' }
  /catalog/articles/{articleId}:
    get:
      summary: Leer artículo publicado (FR-011)
      parameters:
        - { name: articleId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Artículo, content: { application/json: { schema: { $ref: '#/components/schemas/Article' } } } }
        '404': { description: No encontrado }
  /quizzes/{quizId}/attempts:
    post:
      summary: Enviar cuestionario (inicia Saga de calificación → progreso → notificar → auditar)
      parameters:
        - { name: quizId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                  additionalProperties: { type: string }
      responses:
        '201':
          description: Intento calificado y registrado (FR-012/FR-016)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizGradeResult' }

  # --- Editorial (roles editor / coordinador, FR-006..FR-008) ---
...
  /editorial/versions/{versionId}/submit:
    post:
      summary: Enviar a revisión (rol editor)
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: En revisión }
  /editorial/versions/{versionId}/publish:
    post:
      summary: Aprobar y publicar (rol coordinador editorial; debe diferir del editor creador)
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Publicado }
        '403': { description: El coordinador no puede ser el editor creador (FR-008) }

  # --- Simulador (vía Orquestador para auditoría, D-03) ---
...